+++ FCFS/resize_buffer.h
@@ -6,10 +6,11 @@
 #include <string.h>
 #include <stdlib.h>
 #include <stdio.h>
+#include <stdint.h>
 
 struct resize_buffer {
-    unsigned int last_element_index;
-    unsigned int size;
+    uint64_t last_element_index;
+    uint64_t size;
     char         *string;
 };
 
@@ -19,6 +20,6 @@ void rb_destroy(struct resize_buffer*);
 
 void rb_append(struct resize_buffer* b, const char* str);
 
-void rb_remove_substring(struct resize_buffer*, unsigned int, unsigned int);
+void rb_remove_substring(struct resize_buffer*, uint64_t, uint64_t);
 
 #endif /* RESIZE_BUFFER_H */
